requirements.txt
aiofiles==23.2.1
aiohttp==3.7.4.post0
discord.py==2.3.2
google_api_python_client==2.132.0
openai==0.28.1
Pillow==10.3.0
python-dotenv==1.0.1
Requests==2.32.3
tiktoken==0.7.0
